BJS2-78085 1. исправил LanguageToolService
